Handle modules defined via bridge.json the same way as if defined via the assembly attribute
* Ensure module specified in bridge.json is considered the same way as attribute based modules

* Added dedicated integration tests for #3335

* Integration Tests: implement project compare test.

This way of checking projects output consists in grouping a series of
projects in an integration tests' folder, and checking whether they all
output the exact same code. This is useful to check whether a same setting
done in different ways with Bridge produce the exact same result.

* Adds integration tests for the Module setting.

The integration test uses the CompareProjects() mechanism to check
whether both projects produce the same output, where one of the project
sets the module name via its AssemblyInfo.cs and the other sets the
module name via the bridge.json file.

* Erases content from Integration tests 7's bridge.js.

The file contents are not relevant to the test and it just consumes
space.

* Updates integration test 07 to match output when all tests run.

For some reason the output of issue 2806 integration test may switch
between '$module2' and '$module3'. Currently outputs '$module2' if run
alone and '$module3' if all tests are run together. As all tests are run
at once during CI, we refreshed the expected output and added a note to
the source file.