Fixes time zone / daylight saving issue with DateTime.getDayOfYear.
This commit fixes an issue with the client-side implementation of
System.DateTime.getDayOfYear(date), with a faster approach to calculate
the value.

As the solution relies in calculating the leap year, the corresponding
System.DateTime.getIsLeapYear(year) client side implementation was also
optimized to a ~1000 times faster approach.

The performance of getDayOfYear() was of about 30-40 times for better.
- getDayOfYear() benchmark: https://jsperf.com/get-day-of-year
- getIsLeapYear() benchmark: https://jsperf.com/leapyear-calculation