Disables translator parallel files parsing if log level is 'trace'.
This change makes it so the rewriting during
Translator.InspectAssembly.Rewrite() and file parsing during
Translator.InspectAssembly.BuildSyntaxTree() is run serially instead of
in parallel if and only if logger level is set to 'Trace'.

A message stating the parse is going serial is output if so in both
loops.

This improves debugging Bridge in situations reported in issues:
- bridgedotnet/Bridge#3550
- bridgedotnet/Bridge#3557