Adds cultureInfo support to String's StartsWith() and EndsWith()
This adds support just by encapsulating the call into String.equals, forwarding the culture info parameter through.

* [3593] StartsWith method does not compare strings correctly

* Sync .js files

* Refreshes output files.

* Fixes (Starts/Ends)With when string has special characters.

The compared string no longer needs to be escaped as the check up no
longer uses regexp matching syntax. Besides, the escape characters added
to the prefix/suffix checked against were taining the length of the
string to be cropped and checked against.

* Adds unit tests for additional issue 3593 aspects.

The unit tests checks against the side effect imposed by  the initial
implementation of the fix for this issue. Also, it helps identify if a
particular character of the long string is the cause of the issue, as it
iterates through all starts and ends with characters of the full string.

* Refreshes output files.