Improves NUnit integration tests on project 08.
This sets up the combined js output files as presence-only checking, as
it merges in the Bridge.js file and adds a huge unnecessary footprint to
the history. The report file is still checked, and it will always report
size differences, indicating the Bridge.js file changed, thus the
combined scripts that would potentially need checking.