Reviews NUnit test case for issue 697 due to issue 1468.
Issue 1468 introduced a change in which the code used for issue 697's
test case no longer worked. It explored a "general" usage of .Concat()
that is not supported by native .NET implementation, thus it stopped
working.

As the code portion was not vital to the test case itself, it was then
replaced to compilable code using loops and generic collections. The
important part was congruency between minified output and script-write
references.