Increases attempts limit in issue 1509 tests.
There seems to be some times that it won't return a non-integer value in
a big sample of attempts. This would try yet 10x more, until a number with
decimal digits is found. So the likeliness to fail is even smaller.

This test should probably be removed in the future, as this is not
something uniform throughout browsers and not reliable even in browsers
that implement it...