Relaxes threshold of test 3930 for firefox.
The browser's ratio alloc:resize on firefox goes wild as the testing
host performance changes. Adjusted to values that would work in
saucelabs' hosts.

The actual affected browser by this issue is chrome, where the number
was ~10x slower alloc:resize. After the fix, although chrome stood in a
~1:1.5 ratio, Firefox improved the end value (generally faster) yet
locally it gets almost a 1:2 ratio, whereas in saucelabs the value can
go up to around 4.5x times slower.

As the performance is not that really bad, we just adjusted the
threshold, keeping chrome's as reference (for any other browser if
subject to the tests).